add pageCard
