+++ src/pages/imageCardPage/ImageCardPage.module.scss
@@ -1,0 +1,7 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: white;
+}
