+++ src/pages/index.ts
@@ -2,10 +2,12 @@ import Main from './main/main';
 import YandexAuth from './yandexAuth/yandexAuth';
 import Profile from './profile/Profile';
 import NotFound from './notFound/notFound';
+import ImageCardPage from './imageCardPage/ImageCardPage';
 
 export {
   Main,
   YandexAuth,
   Profile,
   NotFound,
+  ImageCardPage,
 };
